+++ app/src/main/java/org/listenbrainz/android/service/BrainzPlayerServiceConnection.kt
@@ -17,7 +17,7 @@ import org.listenbrainz.android.BuildConfig
 import org.listenbrainz.android.model.PlayingTrack.Companion.toPlayingTrack
 import org.listenbrainz.android.model.RepeatMode
 import org.listenbrainz.android.repository.preferences.AppPreferences
-import org.listenbrainz.android.util.BrainzPlayerExtensions.isPlaying
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.isPlaying
 import org.listenbrainz.android.util.ListenSubmissionState
 import org.listenbrainz.android.util.Resource
