+++ app/src/main/java/org/listenbrainz/android/util/brainzplayer/LocalMusicSource.kt
@@ -1,4 +1,4 @@
-package org.listenbrainz.android.util
+package org.listenbrainz.android.util.brainzplayer
 
 import android.support.v4.media.MediaBrowserCompat
 import android.support.v4.media.MediaDescriptionCompat
