+++ app/src/main/java/org/listenbrainz/android/viewmodel/BrainzPlayerViewModel.kt
@@ -27,12 +27,12 @@ import org.listenbrainz.android.repository.brainzplayer.SongRepository
 import org.listenbrainz.android.repository.preferences.AppPreferences
 import org.listenbrainz.android.service.BrainzPlayerService
 import org.listenbrainz.android.service.BrainzPlayerServiceConnection
-import org.listenbrainz.android.util.BrainzPlayerExtensions.currentPlaybackPosition
-import org.listenbrainz.android.util.BrainzPlayerExtensions.isPlayEnabled
-import org.listenbrainz.android.util.BrainzPlayerExtensions.isPlaying
-import org.listenbrainz.android.util.BrainzPlayerExtensions.isPrepared
-import org.listenbrainz.android.util.BrainzPlayerExtensions.toSong
-import org.listenbrainz.android.util.BrainzPlayerUtils.MEDIA_ROOT_ID
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.currentPlaybackPosition
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.isPlayEnabled
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.isPlaying
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.isPrepared
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerExtensions.toSong
+import org.listenbrainz.android.util.brainzplayer.BrainzPlayerUtils.MEDIA_ROOT_ID
 import org.listenbrainz.android.util.Resource
 import javax.inject.Inject
